Fix JSON serialization error - convert numpy types to Python float

+++ train_high_res.py
@@ -123,10 +123,10 @@ def validate(model, dataloader, criterion, device):
     rmse = np.sqrt(np.mean((predictions - targets) ** 2))
 
     return {
-        'loss': total_loss / len(dataloader),
-        'srcc': srcc,
-        'plcc': plcc,
-        'rmse': rmse
+        'loss': float(total_loss / len(dataloader)),
+        'srcc': float(srcc),
+        'plcc': float(plcc),
+        'rmse': float(rmse)
     }
